# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: RestoSmart Restaurant Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List all shifts
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create a shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftBody"
      responses:
        "201":
          description: Created shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
        startTime:
          type: string
        endTime:
          type: string
      required:
        - id
        - employeeId
        - employeeName
        - dayOfWeek
        - startTime
        - endTime
    CreateShiftBody:
      type: object
      properties:
        employeeId:
          type: integer
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
        startTime:
          type: string
        endTime:
          type: string
      required:
        - employeeId
        - dayOfWeek
        - startTime
        - endTime
...